archive-doc-gen/infra/scripts/index_scripts/requirements.txt
azure-storage-file-datalake==12.20.0
azure-ai-inference==1.0.0b9
pypdf==5.6.0
# pyodbc
tiktoken==0.9.0
msal[broker]==1.32.3
azure-identity==1.23.0
azure-ai-textanalytics==5.3.0
azure-search-documents==11.6.0b12
azure-keyvault-secrets==4.9.0
pandas==2.3.0
# pymssql
datetime